+++ app/src/main/java/dev/mkeeda/planetsimulator/model/SimulationPreset.kt
@@ -3,6 +3,6 @@ package dev.mkeeda.planetsimulator.model
 data class SimulationPreset(
     val name: String,
     val description: String,
-    val gravityConstant: Double,
+    val gravityConstant: Float,
     val bodies: List<CelestialBody>
 )
