+++ app/src/main/java/dev/mkeeda/planetsimulator/util/CoordinateConverter.kt
@@ -13,19 +13,19 @@ class CoordinateConverter(
     private val offsetX = (canvasSize.width - SIMULATION_WIDTH * scale) / 2
     private val offsetY = (canvasSize.height - SIMULATION_HEIGHT * scale) / 2
 
-    fun simToScreen(x: Double, y: Double): Offset {
+    fun simToScreen(x: Float, y: Float): Offset {
         return Offset(
-            x = (x * scale + offsetX).toFloat(),
-            y = (y * scale + offsetY).toFloat()
+            x = x * scale + offsetX,
+            y = y * scale + offsetY
         )
     }
 
-    fun scaleToScreen(value: Double): Float {
-        return (value * scale).toFloat()
+    fun scaleToScreen(value: Float): Float {
+        return value * scale
     }
 
     companion object {
-        const val SIMULATION_WIDTH = 1000
-        const val SIMULATION_HEIGHT = 1000
+        const val SIMULATION_WIDTH = 1000f
+        const val SIMULATION_HEIGHT = 1000f
     }
 }
